Make zmanager abstract
